qcom-dload-mode: Don't allow writing to download_mode parameter
Currently regardless of the value written by the user
it would try to enable download mode.

The proper setup is only done while probing the driver
to conditionally set the dump mode.

On top of that the init code disables SDI ( secure debug
watchdog ) if the parameter is set to false but as far as
we can tell there's no way to enable the watchdog back
( as required for proper download mode apparently ).

The solution is to only allow the configuration via
cmdline and exposing the field as read only.

Change-Id: If140c2c886992d475c8d8c93296cb6ded3a878d7
Signed-off-by: Luca Stefani <luca@osomprivacy.com>